refactor: DAO 분리 및 시간값 데이터베이스 값으로 자동 삽입 적용
