+++ src/main/java/racingcar/dao/RacingCarGameDao.java
@@ -1,0 +1,8 @@
+package racingcar.dao;
+
+import racingcar.domain.RacingCars;
+import racingcar.domain.TryCount;
+
+public interface RacingCarGameDao {
+    int insertGame(final RacingCars racingCars, final TryCount tryCount);
+}
