+++ src/main/java/racingcar/dao/RacingCarPlayerDao.java
@@ -1,0 +1,7 @@
+package racingcar.dao;
+
+import racingcar.domain.RacingCars;
+
+public interface RacingCarPlayerDao {
+    void insertGameLog(final RacingCars racingCars, final int gameId);
+}
