+++ src/main/resources/data.sql
@@ -7,7 +7,7 @@ CREATE TABLE PLAY_RESULT
     id          INT NOT NULL AUTO_INCREMENT,
     winners     VARCHAR(50) NOT NULL,
     trial_count INT         NOT NULL,
-    created_at  DATETIME    NOT NULL,
+    created_at  DATETIME    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     PRIMARY KEY (id)
 );
